Move resource state manipulation to `resource/state.rs`
